requirements.txt
jupyter==1.0.0
pytest==6.2.5
pytest-parametrization==2019.1.4
mock==4.0.3
pytest-tstcls==2020.1.1
